pyproject.toml
[project]
name = "python-project-template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.2.8",
    "langchain>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]

[tool.mypy]
strict = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "TRY", "EM", "FA", "SIM108", "S101", "RET505", "TCH002"]
